Finished the biomes, added palm trees, added very basic lights, fixed the normal trees to looks a little better.

+++ src/Acid/World/Block/BlockDatabase.h
@@ -20,7 +20,7 @@ namespace acid
 
         TextureAtlas textureAtlas;
     private:
-        std::array<std::unique_ptr<BlockType>, static_cast<size_t>(BlockID::NUM_TYPES)> _blocks;
+        std::array<std::unique_ptr<BlockType>, static_cast<unsigned>(BlockID::NUM_TYPES)> _blocks;
     };
 }
